XMLParserTest.java

import static org.junit.Assert.*;
import org.junit.Before;
import org.junit.Test;
import java.util.List;

public class XMLParserTest {

    private XMLParser parser;

    @Before
    public void setUp() {
        parser = new XMLParser();
    }

    @Test
    public void testParseValidXML() {
        // Assuming a sample XML file "test_data.xml" with valid credit card data
        List<CreditCard> cards = parser.parseInputFile("input_file.xml");
        assertNotNull(cards);
        assertFalse(cards.isEmpty());

        // Further assertions can be added as required
    }

    // Additional tests for different scenarios can be implemented here
}
